Fix program binaries to use only consistent sizes.
Loading program binaries across different architectures would cause
runtime errors due to use of size_t.

Also fix the ANGLE major and minor version that were clobbered in an
earlier commit.

BUG=angle:647
BUG=371435

+++ src/libGLESv2/ProgramBinary.cpp
@@ -1125,7 +1125,7 @@ bool ProgramBinary::load(InfoLog &infoLog, const void *binary, GLsizei length)
     stream.read(&mUsesPointSize);
     stream.read(&mShaderVersion);
 
-    size_t size;
+    unsigned int size;
     stream.read(&size);
     if (stream.error())
     {
@@ -1192,7 +1192,7 @@ bool ProgramBinary::load(InfoLog &infoLog, const void *binary, GLsizei length)
         stream.read(&uniformBlock.psRegisterIndex);
         stream.read(&uniformBlock.vsRegisterIndex);
 
-        size_t numMembers;
+        unsigned int numMembers;
         stream.read(&numMembers);
         uniformBlock.memberUniformIndexes.resize(numMembers);
         for (unsigned int blockMemberIndex = 0; blockMemberIndex < numMembers; blockMemberIndex++)
